// apps/web/lib/api/models/getMessages200MessagesItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * API
 * API documentation
 * OpenAPI spec version: 1.0.0
 */
import type { GetMessages200MessagesItemChannel } from './getMessages200MessagesItemChannel';
import type { GetMessages200MessagesItemResponseMetadata } from './getMessages200MessagesItemResponseMetadata';
import type { GetMessages200MessagesItemToolCallsItem } from './getMessages200MessagesItemToolCallsItem';
import type { GetMessages200MessagesItemType } from './getMessages200MessagesItemType';

export type GetMessages200MessagesItem = {
  id: string;
  /** @nullable */
  name: string | null;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$ */
  createdAt: string;
  conversationId: string;
  type: GetMessages200MessagesItemType;
  content: string;
  responseMetadata: GetMessages200MessagesItemResponseMetadata;
  /** @nullable */
  toolCalls: GetMessages200MessagesItemToolCallsItem[] | null;
  invalidToolCalls: unknown | null;
  /** @nullable */
  toolCallId: string | null;
  channel: GetMessages200MessagesItemChannel;
};